Added the global variables need for the calculations.
Made a for loop(with C style indexing) that calculates the h_eff based on
"m" matrix. arr_index variable used to point to matlab's indices. Some
tricks made also inside the 'if(arr_index) == 1' to emulate the numpy's
repeat function.

The tests so far seem to be legit. More testing needed here, with more
shapes that include one dismension.