Removed the commands that load the default '.mat. for the demag tensors.
Defined the matrix 'm' as global.

Removed the 'm' argument from the "llg" function call.

Changed the 'alpha' to 0.02 from 0.2.

Fixed the h_zee field matrix to contain the right values at the right
places with a triple for loop to iterate all the dimensions.

Removed one from the for loop that outputs the file to be in index.

For the output format, made the precision a little bit higher.

Added a beep sound at the end, when the program stops.